chore : package move to separate test case
