+++ src/test/java/validation/ValidationTest.java
@@ -1,4 +1,4 @@
-package study;
+package validation;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
